Migrate debug_* from naev.c to a dedicated file.
