+++ src/debug.h
@@ -1,0 +1,15 @@
+/*
+ * See Licensing and Copyright notice in naev.h
+ */
+
+
+#ifndef DEBUG_H
+#  define DEBUG_H
+
+
+void debug_sigInit (void);
+void debug_sigClose (void);
+
+
+#endif /* DEBUG_H */
+
