+++ src/meson.build
@@ -17,6 +17,7 @@ source = files(
    'console.c',
    'damagetype.c',
    'debris.c',
+   'debug.c',
    'dev.c',
    'dev_mapedit.c',
    'dev_outfit.c',
@@ -198,6 +199,7 @@ headers = files(
    'console.h',
    'damagetype.h',
    'debris.h',
+   'debug.h',
    'dev.h',
    'dev_mapedit.h',
    'dev_outfit.h',
